ppc: microwatt: Update for microwatt with standard UART
This partially reverts f03d030080ccf

powerpc/uart: Choose which UART to use at build time, not runtime

And brings back the assumption that 0 in r3 means microwatt. The
runtime detection will be improved subsequently using the device-tree.

When using standalone microwatt, we now use the syscon registers to
detect the UART type and clock. We also get the right frequency for
the processor timebase.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>